Move server logic from main.py to server.py
